Refactor shared scientific operations for 11C and 15C
